XPCError now Codable, returned across connection
Instead of wrapping all errors thrown by the server as `XPCError.remote`, now all of the errors are returned as their own true types. `XPCError.remote` now no longer exists.

The tradeoff in doing this is that the errors which wrapped around error (namely `.encodingError`, `.decodingError`, and `.other`) all now have a String describing that error instead.

For the encoding and decoding errors this isn't much of a tradeoff at all, but for `other` this may turn out to be too much of a tradeoff on the server side where the this can represent the error generated by a function/closure registered with a route. If in practice that turns out to be true, I think a good solution is probably to add a second error callback that's only called for errors thrown by registered route handlers.