Resolves issue #15 — server factory functions now return same instance
XPCMachServer now internally manages a cache of servers and returns an existing one when possible or throws an error if the requirements do not match. Adds "XPCServer Creation.swift" to test this behavior.

All language for the server as well as the client now refers to "retrieve" or "retrieving" instead of "create" or "creating".