Fixes a racy issue exposed by `testShutdownServer`
It turns out occassionally `ServerTerminationIntegrationTest`'s `testShutdownServert` would result in a segfault inside of the finalizer for the connection. I'm pretty sure the issue was that a retained count needs to be taken for the WeakConnection instance (what we don't want to do is add a retain count to the connection itself).