Makes `XPCServerEndpoint` conform to `Codable`
Encoding and decoding "cheats" by using specific functions custom to `XPCEncoderImpl` and `XPCDecoderImpl`. As such if any other encoder or decoder tried to make use of these an error will be thrown.

Added two simple roundtrip tests to verify encoding and decoding works as expected for the anonymous server. However, this is a pretty limited test as it doesn't actually send the endpoint over an XPC connection.