Proper server deallocation, fixes issue #54
`XPCAnonymousServer` and `XPCMachServer` now both start their listener connections in their initializers. However, now when incoming connections are received they don't unconditionally configure and start them. If the server has not been started, the connections are added to a pending array. Once the server is started, those pending connections are configured and started.

Summary of changes:
 - Consolidated all client connection to `XPCServer.startClientConnection(...)` which all three subclasses now call
 - Each instance of `XPCAnonymousServer` and `XPCMachServer` now has its own dispatch queue that it uses for both receiving connections as well as starting incoming connections, this should prevent race conditions. `XPCServer.targetQueue` now exclusively applies to running the handlers associated with the incoming client requests which makes more sense (and is actually what our own documentation says happens).
 - Added a seemingly trivial test which now passes, but without these changes always resulted in a SIGILL.