Optional completion handler `send` and `sendMessage` without reply
Fixes #17

All `send` and `sendMessage` functions now take in a `XPCResponseHandler` (renamed from `XPCReplyHandler`). For those that require a reply, the response handler is non-optional. For those that can't have a reply, they can optionally provide a `XPCResponseHandler<Void>` which will get called upon successful completion or be provided an `XPCError`.

In all cases, none of the functions now `throw` - all errors generated are provided to their handlers (or ignored if no handler is provided).

Two existing tests were modified and two new tests have been added.

This has been designed in a way which should make it relatively straightforward to address #19.