Misc improvements throughout the code base
I realized they're actually called "XPC services" not "XPC Services", so have updated that throughout the comments (both public and internal)

The rest of the the changes are all misc code changes that I think improve things for the better. There's one small public facing change to `XPCError.misconfiguredXPCService` which now has a String to provide a descriptive error message.