Adds `XPCRequestContext` which exposes client info to the server
The core addition here is the introduction of the entirely optional `XPCRequestContext` which has static properties that are accessible within the execution context of a closure called by `XPCServer`.

The motivation behind this PR is that there are times where I need to differ the behavior of the server based on the caller. Since I trust all of my clients I could have them explicitly provide this information, but it seemed more robust to have this be determined client side.

The design is a bit "magic" so I'm torn on it. I considered three different approaches and ultimately settled on this one. The other two considered were:
1. Closures registered with the server *must* have `XPCRequestContext` as a parameter.
2. Closures registered with the server *may* have `XPCRequestContext` as a parameter.

Approach 1 felt far too heavy handed as I don't think the average user of SecureXPC will need to make use of `XPCRequestContext`. Approach 2 would require doubling the number of handler registration functions from 8 to 16 along with all of the associated boiler plate. Additionally it'd make `XPCRequestContext` more prominent in the API and that's undesirable as again I think it'll be needed in a small minority of cases.