Introduces optional error handlers for each route
A global error handler remains, but now exists for global errors like a request for route that's not registered or insecure connection.

Decoding of a message type and encoding of a reply type and/or error type are still sent to the global error handler. Ideally those would also be sent to the route specific one. If we like this approach, can take a look at doing that next.