Renames `route` parameter to `toRoute` for `send` functions
This is done to align with Swift's API Guidelines that state, "Prefer method and function names that make use sites form grammatical English phrases."

The `toRoute` parameter is always the first or second parameter, which makes it quite important. As the guidelines further state, "It is acceptable for fluency to degrade after the first argument or two when those arguments are not central to the call’s meaning."

There are no functional changes to this PR, just a naming change. Of course this is a breaking change.