Adds `requestID` concept
All requests now have a unique identifier and all replies reference this identifier. This is currently useless as the XPC framework itself makes sure that replies are associated back with their requests.

The reason for adding a `requestID` is to enable server handlers to respond with multiple values per request. From the client perspective this will be exposed as an `AsyncSequence` (and some sort of backwards compatibility closure-based version as well). These responses will in fact be sends from the server to the client, containing the `requestID` which then allows for them to be properly associated.

Since this is a fully separable (although by itself useless) part, I thought it best for it to have its own PR first.