Adds a convenience server creation function for login items
This is conceptually similar to the one for executables installed with `SMJobBless`, but instead is for bundles enabled with `SMLoginItemSetEnabled`.

The current message acceptor logic is a work in progress, it needs to be expanded to support connections from anything within the main app bundle. This is to support scenarios such as a Finder Sync Extension wanting to communicate with it: https://developer.apple.com/forums/thread/112628.