Partial support for macOS 13's new SMAppService
 - Adds support for daemons and this has been tested
 - Updates documentation for login items, which _may_ just work, but it's actually unclear because Apple's documentation right now is obviously incorrect. See https://developer.apple.com/forums/thread/709171#709171021
 - No support yet for launch agents

However, this is making the API surface increasing unwieldy so I have thoughts on how to refactor this to in most cases auto-detect which type of server  (and for that matter client as well) and let users explicitly specify a "type" if so desired. But first, wanted to land this change in the current style (although likely not make a release containing it).