Adds an `endpoint` to `XPCServiceServer`
Now all `XPCServer` subclasses provide a server, so it can be part of the class itself which plays nicely with the `forThisProcess(...)` auto-detecting function.

This has been manually tested and works as expected; however, this can't ship as-is because all messages are always being accepted which is no longer safe since the endpoint can be passed around. So something will need to be done to adjust how its `MessageAcceptor` works.