Adds namespacing to `XPCRoute`, adds `serverIdentity` to `XPCClient`
Routes are now namespace allowing for the framework to have its own routes that cannot conflict with routes provided by end users of the API.

This functionality is then used to introduce a noop route which the client calls in order to obtain the identity of the server. This is essentially the client's equivalent of `XPCRequestContext`.

The intention is to build upon this PR with another one that allows the client to enforce a security requirement on the server.