Introduces shared memory capabilities
The `SharedTrivial` property wrapper conforms to `Codable` but is not serializing its boxed value, it's actually mapping it into shared memory which can be accessed and modified across processes. My intention is to follow this up with a to-be-written PR for a collection like type structure (think a fixed size array of trivial types and/or a fixed size `Data` type).

Note - I didn't write any tests for this because they unfortunately don't seem very useful as our tests all run in the same process. I could write some very basic ones, but they're not going to catch actual cross-process issues. I did test this a fair bit using an app with an XPC service.