Namespaces 4 structs and classes under `XPCClient` and `XPCServer`
This performs the following renames/re-namespacings:
- `XPCServerIdentity` -> `XPCClient.ServerIdentity`
- `XPCRequestContext` -> `XPCServer.RequestContext`
- `XPCClientRequirement` -> `XPCServer.ClientRequirement`
- `XPCServerRequirement` -> `XPCClient.ServerRequirement`

The motivation behind this is to take these likely lesser used classes/structs which are only associated with the either `XPCClient` or `XPCServer` and to namespace them within them to make it clearer what their relationship is to their (now) parent class.

No functionality changes and all tests pass.