Interrupted connects finish active sequences, addresses discussion #122
For a given `XPCClient` all active async sequences are terminated with an error if the connection with the server is interrupted. This is the correct behavior because no more responses can possibly be received for this given sequence (although a new requests could be made and might succeed if the server is successfully started once again).