Addresses race conditions during process termination
It's now possible to send a sequential reply which waits on the client to handle it. This allows for a server to terminate its process after knowing the client has received the entire sequence. That said, doing that is not a good approach (this is now documented). However, what this uncovered is this provides a valuable means by which the client can apply back pressure on the server.

This is a breaking change in an async context as there are now new overloaded function names.

Several tests were run in loops 100 times over to ensure the tests reliably pass and are not racing. However, this causes the tests to take a lot longer to complete so they're just running once as reflected in this commit.